pyproject.toml
[tool.poetry]
name = "one-calendar"
version = "0.1.0"
description = ""
authors = ["Morgan C. Nicholson <nich.morgan@outlook.com>"]
readme = "README.md"
packages = [{include = "one_calendar"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.1"
python-dateutil = "^2.8.2"
lxml = "^4.9.2"
antidote = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
